+++ typescript/factory/exercise/social-media-auth-factory.ts
@@ -1,6 +1,6 @@
 import { Authentication } from "./Authentication";
-import { AuthenticationFactory } from "./AuthenticationFactory";
-import { SocialMediaAuthentication } from "./SocialMediaAuthentication";
+import { AuthenticationFactory } from "./authentication-factory";
+import { SocialMediaAuthentication } from "./social-media-authentication";
 
 export class SocialMediaAuthFactory implements AuthenticationFactory {
     createAuthentication(): Authentication {
