+++ typescript/factory/exercise/two-factor-auth-factory.ts
@@ -1,6 +1,6 @@
 import { Authentication } from "./Authentication";
-import { AuthenticationFactory } from "./AuthenticationFactory";
-import { TwoFactorAuthentication } from "./TwoFactorAuthentication";
+import { AuthenticationFactory } from "./authentication-factory";
+import { TwoFactorAuthentication } from "./two-factor-authentication";
 
 export class TwoFactorAuthFactory implements AuthenticationFactory {
     createAuthentication(): Authentication {
