+++ typescript/solid-principles/Paypal-payment-strategy.ts
@@ -1,4 +1,4 @@
-import { PaymentStrategy } from "./PaymentStrategy";
+import { PaymentStrategy } from "./payment-strategy";
 
 export class PayPalPaymentStrategy implements PaymentStrategy {
     processPayment(amount: number) {
