+++ typescript/solid-principles/master-card-payment-strategy.ts
@@ -1,4 +1,4 @@
-import { PaymentStrategy } from "./PaymentStrategy";
+import { PaymentStrategy } from "./payment-strategy";
 
 export class MasterCardPaymentStrategy implements PaymentStrategy {
     processPayment(amount: number) {
