+++ typescript/solid-principles/visa-paymentStrategy.ts
@@ -1,8 +1,7 @@
-import { PaymentStrategy } from "./PaymentStrategy";
+import { PaymentStrategy } from "./payment-strategy";
 
 export class VisaPaymentStrategy implements PaymentStrategy {
-    processPayment(amount: number) {
-        console.log("Processig VISA payment.");
-    }
-    
+  processPayment(amount: number) {
+    console.log("Processig VISA payment.");
+  }
 }
